Improve rendering times of HTML::Builder using some tips from https://crystal-lang.org/reference/1.13/guides/performance.html#avoiding-memory-allocations.
- Uses String::Builder instead IO::Memory
- Don't create intermediate strings (eg. `str << "<" << name` instead of `str << "<#{name}"`

I ran some benchmarks using to_html.cr examples (https://github.com/sbsoftware/to_html.cr/blob/main/benchmark/benchmark.cr).

Results:
```
                           ecr   3.42M (292.20ns) (± 1.63%)  4.27kB/op        fastest
            html_builder (New) 664.95k (  1.50µs) (± 2.18%)  8.39kB/op   5.15× slower
        html_builder (Current) 118.40k (  8.45µs) (± 0.74%)  10.4kB/op  28.90× slower
```

Just New x Current:
```
            html_builder (New) 662.95k (  1.51µs) (± 0.63%)  8.39kB/op        fastest
        html_builder (Current) 119.73k (  8.35µs) (± 0.69%)  10.4kB/op   5.54× slower
```

Ps. I changed `@str` to `@buffer` because I think it's a better name, but I can change it back to `@str`.